refactor(kda): use phase-neutral API module name
Rename the canonical recurrent KDA API and documentation from kda_decode to kda now that one dispatcher serves decode, speculative decode, and prefill. Preserve existing direct imports through a thin compatibility package and update trace identity and Kimi Delta Attention naming.\n\nAI-assisted: Codex
