+++ benchmarks/bench_recurrent_kda.py
@@ -35,7 +35,7 @@ from flashinfer.testing import bench_gpu_time
 
 # Import the recurrent KDA kernel
 try:
-    from flashinfer.kda_decode import _RECURRENT_KDA_AVAILABLE, recurrent_kda
+    from flashinfer.kda import _RECURRENT_KDA_AVAILABLE, recurrent_kda
 
     RECURRENT_KDA_AVAILABLE = _RECURRENT_KDA_AVAILABLE
 except ImportError:
@@ -233,7 +233,7 @@ def run_recurrent_kda_benchmark(args, dtype):
     """Run recurrent KDA decode benchmarks."""
     if not RECURRENT_KDA_AVAILABLE:
         print("Error: recurrent KDA kernel is not available.")
-        print("Make sure flashinfer.kda_decode.recurrent_kda is importable.")
+        print("Make sure flashinfer.kda.recurrent_kda is importable.")
         return
 
     invalid_seq_lens = [t for t in args.seq_len if t < 1]
