+++ tests/trace/fi_trace_out/recurrent_kda_q8_v16_d128.json
@@ -1,9 +1,9 @@
 {
   "name": "recurrent_kda_q8_v16_d128",
-  "description": "Recurrent Key-Driven Attention decode/spec-decode plus exact FlashKDA-compatible ordinary prefill on B200, with per-key-dimension gating and recurrent BF16 state.",
+  "description": "Recurrent Kimi Delta Attention decode/spec-decode plus exact FlashKDA-compatible ordinary prefill on B200, with per-key-dimension gating and recurrent BF16 state.",
   "op_type": "kda",
   "tags": [
-    "fi_api:flashinfer.kda_decode.recurrent_kda",
+    "fi_api:flashinfer.kda.recurrent_kda",
     "stage:decode",
     "stage:prefill",
     "status:verified"
